requirements.txt
fastapi
uvicorn
pillow
easyocr
python-multipart
